refactor(compute-providers): share EC2 runner operations
