+++ lambdas/functions/control-plane/src/scale-runners/types.ts
@@ -3,7 +3,7 @@ export type {
   CreateRunnerResult,
   CreateScaleUpRunnersInput,
   CurrentRunnersInput,
-  LambdaRunnerSource,
+  RunnerSource,
   RunnerLabelResolution,
   RunnerInfo,
   RunnerType,
