+++ lambdas/libs/compute-providers/aws/ec2/src/control-plane/pool.ts
@@ -8,7 +8,7 @@ import type {
   RunnerStatus,
 } from '../../../../core';
 import { createRunners, loadEc2ProviderConfig } from './runner-config';
-import { bootTimeExceeded, listEC2Runners } from './runners';
+import { bootTimeExceeded, listEC2Runners } from '../runners';
 
 const logger = createChildLogger('pool');
 
@@ -46,7 +46,7 @@ async function createEc2PoolRunners(
     numberOfRunners,
     githubInstallationClient,
     createStartRunnerConfig,
-    'pool-lambda',
+    'pool',
   );
   return instances;
 }
